refac: add new hooks to managing input state
