+++ 03-trip/03-03-Next/src/components/Organisms/BoardsNew/index.tsx
@@ -1,48 +1,20 @@
 "use client";
 
-import { useState } from "react";
-import useHooksCreate from "@/commons/hooks/useHooksCreate";
-import { InputField } from "@/components/Molecules/_InputField";
+import { useRouter } from "next/navigation";
+import useSubmitInput from "@/commons/hooks/useSubmitInput";
+import useCreate from "@/commons/hooks/useCreate";
 
+import InputField from "@/components/Molecules/_InputField";
 import AddressField from "@/components/Molecules/_AddrField";
 import ImgField from "@/components/Molecules/_ImgField";
 import Btn from "@/components/Atoms/_Btn";
-import { useRouter } from "next/navigation";
 
 export default function BoardsNewUI() {
-    const [author, setAuthor] = useState("");
-    const [password, setPassword] = useState("");
-    const [title, setTitle] = useState("");
-    const [content, setContent] = useState("");
-    const valid = author && password && title && content;
-
     const Router = useRouter();
 
-    const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-        const id = event.target.id;
-        const value = event.target.value;
-
-        switch (id) {
-            case "author__ID": {
-                setAuthor(value);
-                break;
-            }
-            case "password__ID": {
-                setPassword(value);
-                break;
-            }
-            case "title__ID": {
-                setTitle(value);
-                break;
-            }
-            case "content__ID": {
-                setContent(value);
-                break;
-            }
-        }
-    };
-
-    const onClickCreate = useHooksCreate({ author, password, title, content });
+    const { handleChange, author, password, title, content, valid } =
+        useSubmitInput();
+    const onClickCreate = useCreate({ author, password, title, content });
 
     return (
         <>
